Move out code to different files.
