+++ Cargo.toml
@@ -10,24 +10,23 @@ path = "src/main.rs"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-openssl = { version = '0.10', features = ["vendored"] }
 anyhow = "1.0.65"
 clap = { version = "4.0.17", features = ["derive", "cargo"] }
-envtestkit = "1.1.2"
-humantime = "2.1.0"
+dirs = "4.0.0"
+http-auth-basic = "0.3.3"
 indicatif = "0.17.2"
 log = { version = "0.4.17", features = ["release_max_level_info", "max_level_trace"] }
+openssl = { version = '0.10', features = ["vendored"] }
 prettytable-rs = "0.10.0"
 reqwest = { version = "0.11.12", features = ["json", "blocking", "multipart"] }
+rpassword = "7.2.0"
 serde = { version = "1.0.147", features = ["derive"] }
 serde_json = "1.0.87"
 simple_logger = "4.0.0"
 tempdir = "0.3.7"
 term = "0.7.0"
 thiserror = "1.0.37"
-dirs = "4.0.0"
-http-auth-basic = "0.3.3"
-rpassword = "7.2.0"
 
 [dev-dependencies]
 httpmock = "0.6"
+envtestkit = "1.1.2"
