+++ src/commands/screen.rs
@@ -1,7 +1,6 @@
 use crate::authentication::Authentication;
 use crate::commands;
 use crate::commands::{CommandError, Formatter, OutputType};
-use humantime::format_duration;
 use prettytable::{row, Table};
 use reqwest::StatusCode;
 use std::collections::HashMap;
@@ -43,7 +42,7 @@ impl Formatter for Screens {
                             if uptime.eq("N/A") {
                                 uptime.to_owned()
                             } else {
-                                format_duration(Duration::new(
+                                indicatif::HumanDuration(Duration::new(
                                     uptime.parse::<f64>().unwrap_or(0.0) as u64,
                                     0,
                                 ))
@@ -93,7 +92,7 @@ impl ScreenCommand {
     }
 
     pub fn get(&self, id: &str) -> anyhow::Result<Screens, CommandError> {
-        let endpoint = format!("v4/screens?id=eq.{}", id);
+        let endpoint = format!("v4/screens?id=eq.{id}");
 
         Ok(Screens::new(commands::get(
             &self.authentication,
@@ -131,7 +130,7 @@ impl ScreenCommand {
     }
 
     pub fn delete(&self, id: &str) -> anyhow::Result<(), CommandError> {
-        let endpoint = format!("v3/screens/{}/", id);
+        let endpoint = format!("v3/screens/{id}/");
         commands::delete(&self.authentication, &endpoint)
     }
 }
@@ -139,12 +138,12 @@ impl ScreenCommand {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::commands::*;
+
     use httpmock::{Method::GET, MockServer};
 
     use envtestkit::lock::lock_test;
     use envtestkit::set_env;
-    use httpmock::Method::{DELETE, PATCH, POST};
+    use httpmock::Method::{DELETE, POST};
     use std::ffi::OsString;
     use std::fs;
 
@@ -292,13 +291,13 @@ mod tests {
         let screen = Screens::new(serde_json::from_str("[{\"id\":\"017a5104-524b-33d8-8026-9087b59e7eb5\",\"team_id\":\"016343c2-82b8-0000-a121-e30f1035875e\",\"created_at\":\"2021-06-28T05:07:55+00:00\",\"name\":\"Renat's integrated wired NM\",\"is_enabled\":true,\"coords\":[55.22931, 48.90429],\"last_ping\":\"2021-08-25T06:17:20.728+00:00\",\"last_ip\":null,\"local_ip\":\"192.168.1.146\",\"mac\":\"b8:27:eb:d6:83:6f\",\"last_screenshot_time\":\"2021-08-25T06:09:04.399+00:00\",\"uptime\":\"230728.38\",\"load_avg\":\"0.14\",\"signal_strength\":null,\"interface\":\"eth0\",\"debug\":false,\"location\":\"Kamsko-Ust'inskiy rayon, Russia\",\"team\":\"016343c2-82b8-0000-a121-e30f1035875e\",\"timezone\":\"Europe/Moscow\",\"type\":\"hardware\",\"hostname\":\"srly-4shnfrdc5cd2p0p\",\"ws_open\":false,\"status\":\"Offline\",\"last_screenshot\":\"https://us-assets.screenlyapp.com/01CD1W50NR000A28F31W83B1TY/screenshots/01F98G8MJB6FC809MGGYTSWZNN/5267668e6db35498e61b83d4c702dbe8\",\"in_sync\":false,\"software_version\":\"Screenly 2 Player\",\"hardware_version\":\"Raspberry Pi 3B\",\"config\":{\"hdmi_mode\": 34, \"hdmi_boost\": 2, \"hdmi_drive\": 0, \"hdmi_group\": 0, \"verify_ssl\": true, \"audio_output\": \"hdmi\", \"hdmi_timings\": \"\", \"overscan_top\": 0, \"overscan_left\": 0, \"use_composite\": false, \"display_rotate\": 0, \"overscan_right\": 0, \"overscan_scale\": 0, \"overscan_bottom\": 0, \"disable_overscan\": 0, \"shuffle_playlist\": false, \"framebuffer_width\": 0, \"use_composite_pal\": false, \"framebuffer_height\": 0, \"hdmi_force_hotplug\": true, \"use_composite_ntsc\": false, \"hdmi_pixel_encoding\": 0, \"play_history_enabled\": false}}, {\"id\":\"017a5104-524b-33d8-8026-9087b59e7eb6\",\"team_id\":\"016343c2-82b8-0000-a121-e30f1035875d\",\"created_at\":\"2020-06-28T05:07:55+00:00\",\"name\":\"Not Renat's integrated wired NM\",\"is_enabled\":true,\"coords\":[55.22931, 48.90429],\"last_ping\":\"2020-08-25T06:17:20.728+00:00\",\"last_ip\":null,\"local_ip\":\"192.168.1.146\",\"mac\":\"b8:27:eb:d6:83:6f\",\"last_screenshot_time\":\"2021-08-25T06:09:04.399+00:00\",\"uptime\":\"230728.38\",\"load_avg\":\"0.14\",\"signal_strength\":null,\"interface\":\"eth0\",\"debug\":false,\"location\":\"Kamsko-Ust'inskiy rayon, Russia\",\"team\":\"016343c2-82b8-0000-a121-e30f1035875e\",\"timezone\":\"Europe/Moscow\",\"type\":\"hardware\",\"hostname\":\"srly-4shnfrdc5cd2p0p\",\"ws_open\":false,\"status\":\"Offline\",\"last_screenshot\":\"https://us-assets.screenlyapp.com/01CD1W50NR000A28F31W83B1TY/screenshots/01F98G8MJB6FC809MGGYTSWZNN/5267668e6db35498e61b83d4c702dbe8\",\"in_sync\":false,\"software_version\":\"Screenly 2 Player\",\"hardware_version\":\"Raspberry Pi 3B\",\"config\":{\"hdmi_mode\": 34, \"hdmi_boost\": 2, \"hdmi_drive\": 0, \"hdmi_group\": 0, \"verify_ssl\": true, \"audio_output\": \"hdmi\", \"hdmi_timings\": \"\", \"overscan_top\": 0, \"overscan_left\": 0, \"use_composite\": false, \"display_rotate\": 0, \"overscan_right\": 0, \"overscan_scale\": 0, \"overscan_bottom\": 0, \"disable_overscan\": 0, \"shuffle_playlist\": false, \"framebuffer_width\": 0, \"use_composite_pal\": false, \"framebuffer_height\": 0, \"hdmi_force_hotplug\": true, \"use_composite_ntsc\": false, \"hdmi_pixel_encoding\": 0, \"play_history_enabled\": false}}]").unwrap());
         println!("{}", screen.format(OutputType::HumanReadable));
         let expected_output =
-"+--------------------------------------+---------------------------------+------------------+---------+-------------------------------+------------------+\n\
-| Id                                   | Name                            | Hardware Version | In Sync | Last Ping                     | Uptime           |\n\
-+--------------------------------------+---------------------------------+------------------+---------+-------------------------------+------------------+\n\
-| 017a5104-524b-33d8-8026-9087b59e7eb5 | Renat's integrated wired NM     | Raspberry Pi 3B  |   ❌    | 2021-08-25T06:17:20.728+00:00 | 2days 16h 5m 28s |\n\
-+--------------------------------------+---------------------------------+------------------+---------+-------------------------------+------------------+\n\
-| 017a5104-524b-33d8-8026-9087b59e7eb6 | Not Renat's integrated wired NM | Raspberry Pi 3B  |   ❌    | 2020-08-25T06:17:20.728+00:00 | 2days 16h 5m 28s |\n\
-+--------------------------------------+---------------------------------+------------------+---------+-------------------------------+------------------+\n";
+"+--------------------------------------+---------------------------------+------------------+---------+-------------------------------+--------+\n\
+| Id                                   | Name                            | Hardware Version | In Sync | Last Ping                     | Uptime |\n\
++--------------------------------------+---------------------------------+------------------+---------+-------------------------------+--------+\n\
+| 017a5104-524b-33d8-8026-9087b59e7eb5 | Renat's integrated wired NM     | Raspberry Pi 3B  |   ❌    | 2021-08-25T06:17:20.728+00:00 | 3 days |\n\
++--------------------------------------+---------------------------------+------------------+---------+-------------------------------+--------+\n\
+| 017a5104-524b-33d8-8026-9087b59e7eb6 | Not Renat's integrated wired NM | Raspberry Pi 3B  |   ❌    | 2020-08-25T06:17:20.728+00:00 | 3 days |\n\
++--------------------------------------+---------------------------------+------------------+---------+-------------------------------+--------+\n";
 
         assert_eq!(screen.format(OutputType::HumanReadable), expected_output);
     }
